Set dimnames

A more flexible and pipe-friendly version of dimnames<-.

set_dimnames(x, nm, .dim = NULL)

Arguments

x
an array

nm
A list or character vector.

.dim
a character vector or numeric vector or NULL

Value

x, with dimnames

Details

The word "dimnames" is overloaded, in that it can refer to either the names of the array axes (e.g, metric, timestep, channels), or it can refer to the names of entries along a particular axes (e.g., date1, date2, date3, ...). This function can be used to set either one (by passing nm a character vector and specifying .dim accordingly), or both at the same time (by passing a named list to nm).

Examples

x <- array(1:8, 2:4)

# to set axis names, leave .dim=NULL and pass a character vector
dimnames(set_dimnames(x, c("a", "b", "c")))

#> $a
#> NULL
#> 
#> $b
#> NULL
#> 
#> $c
#> NULL
#> 

# to set names along a single axis, specify .dim
dimnames(set_dimnames(x, c("a", "b", "c"), 2))

#> [[1]]
#> NULL
#> 
#> [[2]]
#> [1] "a" "b" "c"
#> 
#> [[3]]
#> NULL
#> 

# to set an axis name and names along the axis, pass a named list
dimnames(set_dimnames(x, list(axis2 = c("a", "b", "c")), 2))

#> [[1]]
#> NULL
#> 
#> $axis2
#> [1] "a" "b" "c"
#> 
#> [[3]]
#> NULL
#> 

dimnames(set_dimnames(x, list(axis2 = c("a", "b", "c"),
                              axis3 = 1:4), .dim = 2:3))

#> [[1]]
#> NULL
#> 
#> $axis2
#> [1] "a" "b" "c"
#> 
#> $axis3
#> [1] "1" "2" "3" "4"
#> 

# if the array already has axis names, those are used when possible
nx <- set_dimnames(x, paste0("axis", 1:3))
dimnames(nx)

#> $axis1
#> NULL
#> 
#> $axis2
#> NULL
#> 
#> $axis3
#> NULL
#> 

dimnames(set_dimnames(nx, list(axis2 = c("x", "y", "z"))))

#> $axis1
#> NULL
#> 
#> $axis2
#> [1] "x" "y" "z"
#> 
#> $axis3
#> NULL
#> 

dimnames(set_dimnames(nx, c("x", "y", "z"), .dim = "axis2"))

#> $axis1
#> NULL
#> 
#> $axis2
#> [1] "x" "y" "z"
#> 
#> $axis3
#> NULL
#> 

# pass NULL to drop all dimnames, or just names along a single dimension
nx2 <- set_dimnames(nx, c("x", "y", "z"), .dim = "axis2")
dimnames(nx2)

#> $axis1
#> NULL
#> 
#> $axis2
#> [1] "x" "y" "z"
#> 
#> $axis3
#> NULL
#> 

dimnames(set_dimnames(nx2, NULL))

#> NULL

dimnames(set_dimnames(nx2, NULL, 2L))

#> $axis1
#> NULL
#> 
#> $axis3
#> NULL
#> 
#> [[3]]
#> NULL
#>